Conclusão dos testes com composição

+++ jnosql/document/src/main/java/model/Motor.java
@@ -28,7 +28,7 @@ public class Motor {
         return Power;
     }
 
-    public void setPower(float power) {
+    public void setPower(double power) {
         Power = power;
     }
 }
